fix chroot and default user behavior
Use of chroot causes complications when a root is specified; chroot requires
/etc/^(passwd group) and friends be copied to the new root. This is
uneccessary, and is obviated by use of a new function named rootpath.

Files created with a default user are owned by the user of the process rather
than the default user. This is addressed by issuing a chown after create to
reinstate proper permission.